+++ plugins/CMakeLists.txt
@@ -9,7 +9,7 @@ add_subdirectory(producer_api_plugin)
 add_subdirectory(history_plugin)
 add_subdirectory(history_api_plugin)
 add_subdirectory(state_history_plugin)
-add_subdirectory(ethropsten_swap_plugin)
+add_subdirectory(eth_swap_plugin)
 
 add_subdirectory(wallet_plugin)
 add_subdirectory(wallet_api_plugin)
